+++ lib/src/location_picker.dart
@@ -9,7 +9,7 @@ import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart' as intl;
 import 'package:latlong2/latlong.dart';
 
-import 'Widgets/copyright_osm_widget.dart';
+import 'widgets/copyright_osm_widget.dart';
 import 'widgets/wide_button.dart';
 import 'classes.dart';
 
@@ -62,6 +62,10 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final double maxZoomLevel;
 
+  /// [maxBounds] : (LatLngBounds?) set default max bounds of the map (default = null)
+  ///
+  final LatLngBounds? maxBounds;
+
   /// [loadingWidget] : (Widget) show custom  widget until the map finish initialization
   ///
   final Widget? loadingWidget;
@@ -98,10 +102,6 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final ButtonStyle? selectLocationButtonStyle;
 
-  /// [selectLocationTextColor] : (Color) change the color of the select Location text
-  ///
-  final Color? selectLocationTextColor;
-
   /// [selectLocationButtonWidth] : (double) change the width of the select Location button
   ///
   final double? selectLocationButtonWidth;
@@ -109,6 +109,10 @@ class FlutterLocationPicker extends StatefulWidget {
   /// [selectLocationButtonHeight] : (double) change the height of the select Location button
   ///
   final double? selectLocationButtonHeight;
+
+  /// [selectedLocationButtonTextstyle] : set the style of the button text (default = TextStyle(fontSize: 20))
+  ///
+  final TextStyle selectedLocationButtonTextstyle;
 
   /// [selectLocationButtonPositionTop] : (double) change the top position of the select Location button (default = null)
   ///
@@ -174,7 +178,7 @@ class FlutterLocationPicker extends StatefulWidget {
   ///
   final Color? locationButtonBackgroundColor;
 
-  /// [markerIcon] : (IconData) change the marker icon of the map (default = Icon(icons.location_on, color: Colors.red, size: 50))
+  /// [markerIcon] : (IconData) change the marker icon of the map (default = Icon(icons.location_on, color: Colors.blue, size: 50))
   ///
   final Widget? markerIcon;
 
@@ -217,12 +221,14 @@ class FlutterLocationPicker extends StatefulWidget {
   const FlutterLocationPicker({
     Key? key,
     required this.onPicked,
+    this.selectedLocationButtonTextstyle = const TextStyle(fontSize: 20),
     this.onError,
     this.initPosition,
     this.stepZoom = 1,
     this.initZoom = 17,
     this.minZoomLevel = 2,
     this.maxZoomLevel = 18.4,
+    this.maxBounds,
     this.urlTemplate = 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
     this.mapLanguage = 'en',
     this.selectLocationButtonText = 'Set Current Location',
@@ -233,7 +239,6 @@ class FlutterLocationPicker extends StatefulWidget {
     this.showSelectLocationButton = true,
     this.showCurrentLocationPointer = true,
     this.selectLocationButtonStyle,
-    this.selectLocationTextColor,
     this.selectLocationButtonWidth,
     this.selectLocationButtonHeight,
     this.selectLocationButtonPositionTop,
@@ -388,27 +393,18 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
   ///   longitude (double): The longitude of the location.
   void setNameCurrentPos(double latitude, double longitude) async {
     var client = http.Client();
-    setState(() {
-      isLoading = true;
-    });
     String url =
         'https://nominatim.openstreetmap.org/reverse?format=json&lat=$latitude&lon=$longitude&zoom=18&addressdetails=1&accept-language=${widget.mapLanguage}';
 
     try {
-      var response = await client.post(Uri.parse(url));
+      var response = await client.get(Uri.parse(url));
       var decodedResponse =
           jsonDecode(utf8.decode(response.bodyBytes)) as Map<dynamic, dynamic>;
       _searchController.text =
-          decodedResponse['display_name'] ?? "Search Location";
-      setState(() {
-        isLoading = false;
-      });
+          decodedResponse['display_name'] ?? "This Location is not accessible";
+      setState(() {});
     } on Exception catch (e) {
       onError(e);
-    } finally {
-      setState(() {
-        isLoading = false;
-      });
     }
   }
 
@@ -423,12 +419,20 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
     var client = http.Client();
     String url =
         'https://nominatim.openstreetmap.org/reverse?format=json&lat=${_mapController.center.latitude}&lon=${_mapController.center.longitude}&zoom=18&addressdetails=1&accept-language=${widget.mapLanguage}';
-
-    var response = await client.post(Uri.parse(url));
+    var response = await client.get(Uri.parse(url));
     var decodedResponse =
         jsonDecode(utf8.decode(response.bodyBytes)) as Map<dynamic, dynamic>;
-    String displayName = decodedResponse['display_name'];
-    return PickedData(center, displayName, decodedResponse['address']);
+    String displayName = "This Location is not accessible";
+    Map<String, dynamic> address;
+
+    if (decodedResponse['display_name'] != null) {
+      displayName = decodedResponse['display_name'];
+      address = decodedResponse['address'];
+    } else {
+      center = const LatLong(0, 0);
+      address = decodedResponse as Map<String, dynamic>;
+    }
+    return PickedData(center, displayName, address);
   }
 
   @override
@@ -560,7 +564,11 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
                       widget.searchbarInputFocusBorderp ?? inputFocusBorder,
                   hintStyle: TextStyle(color: widget.searchBarHintColor),
                   suffixIcon: IconButton(
-                    onPressed: () => _searchController.clear(),
+                    onPressed: () {
+                      _searchController.clear();
+                      _options.clear();
+                      setState(() {});
+                    },
                     icon: Icon(
                       Icons.clear,
                       color: widget.searchBarTextColor,
@@ -577,7 +585,7 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
                     try {
                       String url =
                           'https://nominatim.openstreetmap.org/search?q=$value&format=json&polygon_geojson=1&addressdetails=1&accept-language=${widget.mapLanguage}';
-                      var response = await client.post(Uri.parse(url));
+                      var response = await client.get(Uri.parse(url));
                       var decodedResponse =
                           jsonDecode(utf8.decode(response.bodyBytes))
                               as List<dynamic>;
@@ -645,9 +653,9 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
               heroTag: "btn3",
               backgroundColor: widget.locationButtonBackgroundColor,
               onPressed: () async {
-                setState(() {
-                  isLoading = true;
-                });
+                // setState(() {
+                //   isLoading = true;
+                // });
                 _determinePosition().then((currentPosition) {
                   _animatedMapMove(
                       LatLng(
@@ -669,10 +677,12 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
     return Positioned.fill(
         child: FlutterMap(
       options: MapOptions(
-          center: initPosition,
-          zoom: initPosition != null ? widget.initZoom : widget.minZoomLevel,
-          maxZoom: widget.maxZoomLevel,
-          minZoom: widget.minZoomLevel),
+        center: initPosition,
+        zoom: initPosition != null ? widget.initZoom : widget.minZoomLevel,
+        maxZoom: widget.maxZoomLevel,
+        minZoom: widget.minZoomLevel,
+        maxBounds: widget.maxBounds,
+      ),
       mapController: _mapController,
       children: [
         TileLayer(
@@ -704,7 +714,7 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
           child: widget.markerIcon ??
               const Icon(
                 Icons.location_pin,
-                color: Colors.red,
+                color: Colors.blue,
                 size: 50,
               ),
         ),
@@ -735,7 +745,7 @@ class _FlutterLocationPickerState extends State<FlutterLocationPicker>
                   }));
             },
             style: widget.selectLocationButtonStyle,
-            textColor: widget.selectLocationTextColor,
+            textStyle: widget.selectedLocationButtonTextstyle,
             width: widget.selectLocationButtonWidth,
             height: widget.selectLocationButtonHeight,
           ),