+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.2
+
+* SetLocationError handling
+* Support more customizations for Select Location button Text
+* Remove loading when the current location is changed.
+* Remove SafeArea from map
+* Added maxBounds
+* Update dependencies
+
 ## 1.2.1
 
 * Added OSM copyrights Badge